Added -k flag
If set, this flag prevent the deletion of the generated slnx.config files.
Can be useful if there are such files for example required for testing.

+++ Test/SlnLauncher.Test/AppBaseFileWriter.cs
@@ -18,10 +18,14 @@ namespace SlnLauncher.Test
             return File.Exists(filePath);
         }
 
-        public void DeleteFile(string filePath)
+        public virtual void DeleteFile(string filePath)
         {
             filePath = GetPath(filePath);
-            File.Delete(filePath);
+
+            if (!filePath.Contains("TestSubFolder") || File.Exists(filePath))
+            {
+                File.Delete(filePath);
+            }
         }
 
         public void AppendAllText(string path, string text)
